feat: Implement message handling logic with custom hook and refactor App component
